Add resonance to the noise to produce a brass-like background sound
The noise is necessary to produce the effect of wind blowing through a tube. White noise with a high-pass filter and low-pass filter pretty much sounds like wind. Add a reverberation effect to this and you get a brass like effect as if the wind is blowing through a metal tube.
Perhaps I can get this to sound a bit more wooden if the noise is dampened non-linearly depending on frequency. It seems like it would be appropriate to reduce the high frequencies in the reverberation as if they get absorbed by the wood of the flute.